A media player program to run all your video formats.
A media player program built in python to run all types of video and mp3 formats.
But to run this program,
First off, you’ll need to make sure you have the following:

    Python (duh!)
    wxPython
    mplayer
    MPlayerCtrl

Once you’ve confirmed that you have all that, we can continue. Let’s look at some code!.

A  Screenshot of the media player is added so you can have a look at the screenshot and get a idea of how the media player will look like.